moved paramter execution of import command into method of Import struct
